support running envoy as a local process (#892)
CircleCI does not support mounting volumes to containers, so configuring envoy in a container becomes harder. Instead, we will be running envoy as a local process in CircleCI.

Unfortunately, envoy is not supported in nix on MacOS, so for development we must also support the current docker-based deployment, or require macOS users to install envoy using brew.

This PR therefore adds support for running envoy locally, which will in a future PR be used in CircleCI.

+++ envoy-proxy-dev/stop-envoy.sh
@@ -1,5 +1,16 @@
 #!/usr/bin/env bash
 
-docker stop envoy
+if [ ! -f "envoy.pid" ]; then
+  echo "The file envoy.pid does not exist, not stopping envoy"
+  exit 1
+fi
 
+PID=$(cat envoy.pid)
+if [ "$PID" == "docker" ]; then
+  docker stop envoy
+else
+  kill "$PID"
+fi
+
+rm envoy.pid
 rm envoy-out.json
